四维技术的首次尝试，效果不是很理想，所以先写了个需求文档，方便后面写，目前只实现了四维空间直接向二维空间的投影，摄像头，投影需要的各种矩阵方法。现在决定改底层逻辑代码，也是首次准备使用JPanel这玩意，所以先保存一下。

+++ src/main/java/org/gameStart/testOne/Line.java
@@ -28,10 +28,10 @@ public class Line extends Model {
         g.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON);
         g.setColor(color);
         g.drawLine(
-                point1.getX()-getX()+point1.r,
-                point1.getY()-getY()+point1.r,
-                point2.getX()-getX()+point2.r,
-                point2.getY()-getY()+point2.r
-                );
+                (int) (point1.getX()-getX()+point1.r),
+                (int) (point1.getY()-getY()+point1.r),
+                (int) (point2.getX()-getX()+point2.r),
+                (int) (point2.getY()-getY()+point2.r)
+        );
     }
 }
